refactor(ooda_loop/types): split tests to sibling tests_types.rs (#1266)
- src/ooda_loop/types.rs: 571 → 307 LOC
- New src/ooda_loop/tests_types.rs (261 LOC)

Pure refactor: 17 tests pass.
